hf_space/requirements.txt
gradio>=4.0.0
requests==2.31.0
Pillow>=10.3.0
numpy>=1.26.4
torch>=2.0.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
